anvil_zksync::cli

Struct Cli

pub struct Cli {

    pub command: Option<Command>,
    pub offline: bool,
    pub health_check_endpoint: bool,
    pub config_out: Option<String>,
    pub port: Option<u16>,
    pub host: Vec<IpAddr>,
    pub chain_id: Option<u32>,
    pub show_node_config: Option<bool>,
    pub show_storage_logs: Option<ShowStorageLogs>,
    pub show_vm_details: Option<ShowVMDetails>,
    pub show_gas_details: Option<ShowGasDetails>,
    pub verbosity: u8,
    pub l1_gas_price: Option<u64>,
    pub l2_gas_price: Option<u64>,
    pub l1_pubdata_price: Option<u64>,
    pub price_scale_factor: Option<f64>,
    pub limit_scale_factor: Option<f32>,
    pub override_bytecodes_dir: Option<String>,
    pub enforce_bytecode_compression: Option<bool>,
    pub dev_system_contracts: Option<SystemContractsOptions>,
    pub system_contracts_path: Option<PathBuf>,
    pub protocol_version: Option<ProtocolVersionId>,
    pub evm_interpreter: bool,
    pub zksync_os_group: ZKsyncOsGroup,
    pub log: Option<LogLevel>,
    pub log_file_path: Option<String>,
    pub silent: Option<bool>,
    pub cache: Option<CacheType>,
    pub reset_cache: Option<bool>,
    pub cache_dir: Option<String>,
    pub accounts: u64,
    pub balance: u64,
    pub timestamp: Option<u64>,
    pub init: Option<Genesis>,
    pub state: Option<PathBuf>,
    pub state_interval: Option<u64>,
    pub dump_state: Option<PathBuf>,
    pub preserve_historical_states: bool,
    pub load_state: Option<PathBuf>,
    pub mnemonic: Option<String>,
    pub mnemonic_random: Option<usize>,
    pub mnemonic_seed: Option<u64>,
    pub derivation_path: Option<String>,
    pub auto_impersonate: bool,
    pub block_time: Option<Duration>,
    pub no_mining: bool,
    pub allow_origin: String,
    pub no_cors: bool,
    pub order: TransactionOrder,
    pub l1_group: Option<L1Group>,
    pub auto_execute_l1: Option<bool>,
    pub base_token_symbol: Option<String>,
    pub base_token_ratio: Option<Ratio<u64>>,

}

Fields
command: Option<Command>offline: bool

Run in offline mode (disables all network requests).

health_check_endpoint: bool

Enable health check endpoint. It will be available for GET requests at /health. The endpoint will return 200 OK if the node is healthy.

config_out: Option<String>

Writes output of anvil-zksync as json to user-specified file.

port: Option<u16>

Port to listen on (default: 8011).

host: Vec<IpAddr>

The hosts the server will listen on.

chain_id: Option<u32>

Specify chain ID (default: 260).

show_node_config: Option<bool>

If true, prints node config on startup.

show_storage_logs: Option<ShowStorageLogs>

Show storage log information.

show_vm_details: Option<ShowVMDetails>

Show VM details information.

show_gas_details: Option<ShowGasDetails>

Show gas details information.

verbosity: u8

Increments verbosity each time it is used. (-vv, -vvv)

Example usage:

-vv => verbosity level 2 (includes user calls, user L1-L2 logs, and event calls)

-vvv => level 3 (includes system calls, system L1-L2 logs, and system event calls)

-vvvv => level 4 (includes system calls, system event calls, and precompiles)

-vvvvv => level 5 (includes everything)

l1_gas_price: Option<u64>

Custom L1 gas price (in wei).

l2_gas_price: Option<u64>

Custom L2 gas price (in wei).

l1_pubdata_price: Option<u64>

Custom L1 pubdata price (in wei).

price_scale_factor: Option<f64>

Gas price estimation scale factor.

limit_scale_factor: Option<f32>

Gas limit estimation scale factor.

override_bytecodes_dir: Option<String>

Directory to override bytecodes.

enforce_bytecode_compression: Option<bool>

Enforces bytecode compression (default: false).

dev_system_contracts: Option<SystemContractsOptions>

Option for system contracts (default: built-in).

system_contracts_path: Option<PathBuf>

Override the location of the compiled system contracts.

protocol_version: Option<ProtocolVersionId>

Protocol version to use for new blocks (default: 26). Also affects revision of built-in contracts that will get deployed (if applicable).

evm_interpreter: bool

Enables EVM interpreter.

zksync_os_group: ZKsyncOsGroup

ZKsync OS detailed config.

log: Option<LogLevel>

Log level (default: info).

log_file_path: Option<String>

Log file path (default: anvil-zksync.log).

silent: Option<bool>

If true, the tool will not print anything on startup.

cache: Option<CacheType>

Cache type (none, memory, or disk). Default: “disk”.

reset_cache: Option<bool>

Reset the local disk cache.

cache_dir: Option<String>

Cache directory location for disk cache (default: .cache).

accounts: u64

Number of dev accounts to generate and configure.

balance: u64

The balance of every dev account in Ether.

timestamp: Option<u64>

The timestamp of the genesis block.

init: Option<Genesis>

Initialize the genesis block with the given genesis.json file.

state: Option<PathBuf>

This is an alias for both –load-state and –dump-state.

It initializes the chain with the state and block environment stored at the file, if it exists, and dumps the chain’s state on exit.

state_interval: Option<u64>

Interval in seconds at which the state and block environment is to be dumped to disk.

See –state and –dump-state

dump_state: Option<PathBuf>

Dump the state and block environment of chain on exit to the given file.

If the value is a directory, the state will be written to <VALUE>/state.json.

preserve_historical_states: bool

Preserve historical state snapshots when dumping the state.

This will save the in-memory states of the chain at particular block hashes.

These historical states will be loaded into the memory when --load-state / --state, and aids in RPC calls beyond the block at which state was dumped.

load_state: Option<PathBuf>

Initialize the chain from a previously saved state snapshot.

mnemonic: Option<String>

BIP39 mnemonic phrase used for generating accounts. Cannot be used if mnemonic_random or mnemonic_seed are used.

mnemonic_random: Option<usize>

Automatically generates a BIP39 mnemonic phrase and derives accounts from it. Cannot be used with other mnemonic options. You can specify the number of words you want in the mnemonic. [default: 12]

mnemonic_seed: Option<u64>

Generates a BIP39 mnemonic phrase from a given seed. Cannot be used with other mnemonic options. CAREFUL: This is NOT SAFE and should only be used for testing. Never use the private keys generated in production.

derivation_path: Option<String>

Sets the derivation path of the child key to be derived. [default: m/44’/60’/0’/0/]

auto_impersonate: bool

Enables automatic impersonation on startup. This allows any transaction sender to be simulated as different accounts, which is useful for testing contract behavior.

block_time: Option<Duration>

Block time in seconds for interval sealing. If unset, node seals a new block as soon as there is at least one transaction.

no_mining: bool

Disable auto and interval mining, and mine on demand instead.

allow_origin: String

The cors allow_origin header

no_cors: bool

Disable CORS.

order: TransactionOrder

Transaction ordering in the mempool.

l1_group: Option<L1Group>auto_execute_l1: Option<bool>

Enable automatic execution of L1 batches

base_token_symbol: Option<String>

Base token symbol to use instead of ‘ETH’.

base_token_ratio: Option<Ratio<u64>>

Base token conversion ratio (e.g., ‘40000’, ‘628/17’).

Implementations

impl Cli

pub fn deprecated_config_option()

Checks for deprecated options and warns users.

pub fn into_test_node_config( self, ) -> Result<TestNodeConfig, AnvilEnvironmentError>

Converts the CLI arguments to a TestNodeConfig.

pub fn into_telemetry_props(self) -> TelemetryProps

Converts the CLI arguments to a TelemetryProps to be used as event props.

fn account_generator(&self) -> AccountGenerator